feat: Add QR code generation for sharing
Includes the addition of the `qrcode-generator` library to facilitate QR code creation.
This allows for easy sharing of the application's home page via a scannable QR code.
The QR code generation is implemented in the `PrintOrderView` component and is displayed on the print page.
Also, minor updates to mock data and Swiper type definitions.